requirements.txt
alembic==1.8.0
anyio==3.6.1
apache-airflow==2.3.2
apache-airflow-providers-ftp==2.1.2
apache-airflow-providers-http==2.1.2
apache-airflow-providers-imap==2.2.3
apache-airflow-providers-sqlite==2.1.3
apispec==3.3.2
argcomplete==2.0.0
attrs==20.3.0
Automat==20.2.0
Babel==2.10.1
beautifulsoup4==4.11.1
blinker==1.4
cachelib==0.7.0
cattrs==1.10.0
certifi==2020.12.5
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
connexion==2.13.1
constantly==15.1.0
cron-descriptor==1.2.24
croniter==1.3.5
cryptography==36.0.2
cssselect==1.1.0
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.1.1
distlib==0.3.4
dnspython==2.2.1
docutils==0.18.1
email-validator==1.2.1
filelock==3.6.0
Flask==1.1.2
Flask-AppBuilder==3.4.5
Flask-Babel==2.0.0
Flask-Caching==1.11.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
graphviz==0.20
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
incremental==21.3.0
inflection==0.5.1
itemadapter==0.6.0
itemloaders==1.0.4
itsdangerous==1.1.0
Jinja2==3.0.3
jmespath==1.0.1
jsonschema==4.5.1
lazy-object-proxy==1.7.1
lockfile==0.12.2
lxml==4.9.0
Mako==1.2.0
Markdown==3.3.7
MarkupSafe==2.0.1
marshmallow==3.16.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
numpy==1.22.4
packaging==21.3
pandas==1.4.2
parsel==1.6.0
pathspec==0.9.0
pendulum==2.1.2
platformdirs==2.5.2
pluggy==1.0.0
praw==7.6.0
prawcore==2.3.0
prison==0.2.1
Protego==0.2.1
psutil==5.9.1
psycopg2==2.8.6
py4j==0.10.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
Pygments==2.12.0
PyJWT==1.7.1
pymongo==4.1.1
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
PySocks==1.7.1
pyspark==3.2.1
python-daemon==2.3.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.1
pytzdata==2020.1
PyYAML==5.4.1
queuelib==1.6.2
requests==2.27.1
requests-file==1.5.1
rfc3986==1.5.0
rich==12.4.4
Scrapy==2.6.1
scrapy-splash==0.8.0
service-identity==21.1.0
setproctitle==1.2.3
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.9
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.2
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
tldextract==3.3.0
Twisted==22.4.0
typing_extensions==4.2.0
unicodecsv==0.14.1
update-checker==0.18.0
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
w3lib==1.22.0
websocket-client==1.3.2
Werkzeug==1.0.1
wrapt==1.14.1
WTForms==2.3.3
zipp==3.8.0
zope.interface==5.4.0
